requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.5.4
decorator==5.1.1
distlib==0.3.8
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-appengine-logging==1.4.4
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-functions==1.0.4
google-cloud-logging==3.10.0
google-cloud-pubsub==2.22.0
google-cloud-secret-manager==2.20.1
google-cloud-storage==2.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.12.7
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
lxml==5.2.2
matplotlib-inline==0.1.7
monday==1.2.6
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipe==2.2
pkg_resources==0.0.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyproject-api==1.7.1
PySocks==1.7.1
pytest==7.0.1
pytest-cov==3.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
scraper==0.1.0
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tenacity==8.5.0
tomli==2.0.1
tox==4.16.0
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
